data_model/requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=9.5.0
tqdm>=4.65.0
tensorboard>=2.13.0

# 3D Processing
trimesh>=3.21.0
open3d>=0.17.0

# Optional: FEA Integration
# scipy>=1.10.0
# pyvista>=0.40.0

# Utilities
matplotlib>=3.7.0
opencv-python>=4.7.0
